add pf view

+++ restd.py
@@ -4,18 +4,18 @@ import datetime
 import stat
 from views import app
 
-token = jwt.encode({'username' : 'admin',
+token = jwt.encode({'username' : app.config['LOCAL_API_USER'],
         'exp' : datetime.datetime.utcnow() + datetime.timedelta(days=365)},
         app.config['SECRET_KEY'])
 
 home = os.environ['HOME']
-with open(f"{home}/.restd.token", "w") as file:
-    file.write(token)
+if not os.path.isfile(f"{home}/.restd.token"):
+    with open(f"{home}/.restd.token", "w") as file:
+        file.write(token)
 
-os.chmod(f"{home}/.restd.token", stat.S_IRWXU)
+    os.chmod(f"{home}/.restd.token", stat.S_IRWXU)
 
 def main():
-    print(" * ** Warning this is running in debug mode. For production deployment refer to https://flask.palletsprojects.com/en/1.1.x/deploying/wsgi-standalone/#twisted-web .")
     app.run(host='0.0.0.0',port=app.config['LISTEN_PORT'],debug=True)
 
 if __name__ == '__main__':
